requirements.txt
discord.py
PyNaCl
